added user profile with logs

+++ app/Repositories/ChampionRepository.php
@@ -77,11 +77,28 @@ class ChampionRepository
      */
     public function like(Champion $champion)
     {
+        /**
+         * @var User $user
+         */
         $user = JWTAuth::user();
+
+        /**
+         * @var bool $alreadyLikes
+         */
+        $alreadyLikes = $user->likes()->whereId($champion->id)->exists();
+        if ($alreadyLikes === true) {
+            return;
+        }
 
         // syncWithoutDetaching ensures that user can like a champion only once
         /** @var User $user */
         $user->likes()->syncWithoutDetaching([$champion->id]);
+
+        $logRepository = new LogRepository;
+        $logRepository->createUserLog(
+            UserLog::Like,
+            $champion->name,
+        );
     }
 
     public function dislike(Champion $champion)
@@ -91,8 +108,22 @@ class ChampionRepository
          */
         $user = JWTAuth::user();
 
+        /**
+         * @var bool $alreadyLikes
+         */
+        $alreadyLikes = $user->likes()->whereId($champion->id)->exists();
+        if ($alreadyLikes === false) {
+            return;
+        }
+
         // Remove the association between the user and the champion
         $user->likes()->detach($champion->id);
+
+        $logRepository = new LogRepository;
+        $logRepository->createUserLog(
+            UserLog::Dislike,
+            $champion->name,
+        );
     }
 
     /**
